fix/UpdateCommentsBar
This PR improves the previous implementation of the comments bar.
Firstly, it enables the ability for the # of likes+comments
(which has been moved into the comment bar) updates immediately
when the user likes or submits a comment.
It also updates the comment bar to better represent the figma,
mainly turning the message icon into a "modal" to then enter a comment
into.
Now, if a user would like to see the comments, they can press the
grey "Comments (#)" text.

Please note, this change only exists on the housing page with listings.
There will be a future ticket where this comment bar will be
implemented to all posts.