feature/PullToRefresh
This ticket implements a pull to refresh feature on each page that
has postings. If a new post is created while the user has loaded the
page, they can pull down at the top to call the new posts.

Expo Project to Test Code: https://expo.io/@trentstauffuw/umigrate

(No tests as this is not a component, RefreshControl is a property of FlatList)