feature/PullToRefresh : Cleans up Code
I realized that there was a lot of redudant functions and code.
So, I've moved everything into one simple function which should
make refactoring later on easier.

Also, you'll notice the time it takes to refresh is now longer.
Now, instead of a hardcoded refresh time, the icon will continue to
spin until all of the API calls have been made, making the new listings
appear at the right time when the refreshing stops. This makes the
process longer (only visually) but is more honest with the user.

The time it takes to refresh will shorten with a faster server in
the future.