feature/PullToRefresh (#344)
* feature/PullToRefresh

This ticket implements a pull to refresh feature on each page that
has postings. If a new post is created while the user has loaded the
page, they can pull down at the top to call the new posts.

Expo Project to Test Code: https://expo.io/@trentstauffuw/umigrate

(No tests as this is not a component, RefreshControl is a property of FlatList)

* feature/PullToRefresh

Small Fix

* feature/PullToRefresh : Cleans up Code

I realized that there was a lot of redudant functions and code.
So, I've moved everything into one simple function which should
make refactoring later on easier.

Also, you'll notice the time it takes to refresh is now longer.
Now, instead of a hardcoded refresh time, the icon will continue to
spin until all of the API calls have been made, making the new listings
appear at the right time when the refreshing stops. This makes the
process longer (only visually) but is more honest with the user.

The time it takes to refresh will shorten with a faster server in
the future.

Co-authored-by: Deandre Baker <49622234+deandrebaker@users.noreply.github.com>