WIP: Create Component to Create Community Posts (#364)
* Implement basic community create page

Most of the buttons don't work as of yet, and the form doesn't change depending on which post type is selected yet.

* Add poll option input

* Finish implementing basic form input for all community post types

* Packed all of the text inputs into their own custom component

This is to reduce copy-pasted code.

* Add x button to remove poll options

* Add API calls to polls and posts endpoints and update user info fetching

* Add a component containing common components in create forms

* Add region selector modal

* Move community select modal to its own component and add ability to move between search view to select view

Actual search functionality doesn't work, but you can select options from the search view.

* Reorder options buttons depending on which search option is selected

The selected one gets pushed to the top of the list

* Implement basic community filtering

* Clean up some comments and correct an import

* Disable share button on polls screen when no poll option is selected

* Remove unnecessary comment

* Move components common to all create pages to a dedicated folder

* Update share button disabling logic

Now, the button is disabled when the title is empty, the list of poll options on the poll page is empty, or any one of the poll options is a blank string.

* Remove unused function

* Clean up community search bar text input

I didn't realize that TextInput has a border style prop. I just found out, so I used it.

* Improve search options display

Now it's wrapped in a scrollview, so when more communities are added, we can scroll through them.

* Replace 'region' with 'community'
Also update some comments.

* Implement date and time pickers and button to show them

The date picker appears when you click it, the time picker should appear shortly after. Works on my phone (android), untested on iPhone.

* Improve the picker for iOS

Had to specifically take the case that the phone displaying the picker was iOS

* Add ability to send events to the API

* Remove console.logs and add comments

* Run prettier on a file that was missed

* Update price scale calculations

It now correctly chooses a scale of $ when the price is between 1 and 20 dollars. It chose free before.

* Await the poll option endpoint post function

* Use getUserData()

* Make formatDate an arrow function

* Alter the PollOptionsEndpoint class

No longer extends AbstractEndpoint